+++ CHANGELOG.md
@@ -10,6 +10,9 @@ based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - Begin work on the /awaken command.
 - Add basic database support with PDO.
 - Add ext-uv for a more performant loop.
+- start.php: Bot commands should not be hardcoded.
+- register.php: Bot commands should not be hardcoded.
+- register.php: We need descriptions for each command.
 
 ## [0.0.1] - 2023-??-??